requirements-macos.txt
--find-links https://github.com/cansik/pyrealsense2-macosx/releases
--find-links https://github.com/cansik/data-storage/releases/tag/packages

visiongraph[realsense]==0.1.9
pyopengl

# os-specific
syphonpy
